♻️ refactor / 미들웨어로 페이지 구성
